chore: move glob to react-router
