--- packages/fullstack/examples/glob/README.md
@@ -1,5 +1,0 @@
-# Glob
-
-Building simple framework with file system routing using [`import.meta.glob`](https://vite.dev/guide/features.html#glob-import) and `?assets` import.
-
-[![Open in StackBlitz](https://developer.stackblitz.com/img/open_in_stackblitz.svg)](https://stackblitz.com/github/hi-ogawa/vite-plugins/tree/main/packages/fullstack/examples/glob)
