--- packages/fullstack/examples/glob/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@hiogawa/vite-plugin-fullstack-examples-glob",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "srvx --prod -s ../client dist/ssr/index.js"
-  },
-  "devDependencies": {
-    "@hiogawa/vite-plugin-fullstack": "latest",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
-    "@vitejs/plugin-react": "latest",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-router": "7.9.3",
-    "srvx": "^0.8.7",
-    "vite": "^7.1.5",
-    "vite-plugin-devtools-json": "^1.0.0"
-  }
-}
