--- packages/fullstack/examples/glob/src/routes/about/page.tsx
@@ -1,13 +1,0 @@
-export function Component() {
-  return (
-    <main>
-      <h1>About</h1>
-      <div className="card">
-        <p>
-          This is a simple React Router demo app built with Vite Plugin
-          Fullstack.
-        </p>
-      </div>
-    </main>
-  );
-}
