--- packages/fullstack/examples/glob/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "include": ["src", "*.ts"],
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "erasableSyntaxOnly": true,
-    "allowImportingTsExtensions": true,
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "skipLibCheck": true,
-    "verbatimModuleSyntax": true,
-    "noEmit": true,
-    "moduleResolution": "Bundler",
-    "module": "ESNext",
-    "target": "ESNext",
-    "lib": ["ESNext", "DOM", "DOM.Iterable"],
-    "types": ["vite/client", "@hiogawa/vite-plugin-fullstack/types"]
-  }
-}
