+++ packages/fullstack/examples/react-router/README.md
@@ -1,3 +1,5 @@
-# React Router Custom Framework
+# React Router
+
+This example showcases [`import.meta.glob`](https://vite.dev/guide/features.html#glob-import) and `?assets` import to create a simple framework with file system routing.
 
 [![Open in StackBlitz](https://developer.stackblitz.com/img/open_in_stackblitz.svg)](https://stackblitz.com/github/hi-ogawa/vite-plugins/tree/main/packages/fullstack/examples/react-router)
