+++ packages/fullstack/examples/react-router/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@hiogawa/vite-plugin-fullstack-examples-react-router",
+  "name": "@hiogawa/vite-plugin-fullstack-examples-glob",
   "private": true,
   "type": "module",
   "scripts": {
