+++ packages/fullstack/examples/react-router/src/framework/routes.ts
