+++ packages/fullstack/examples/react-router/src/layout.tsx
