+++ packages/fullstack/examples/react-router/src/routes/about/page.tsx
