+++ packages/fullstack/examples/react-router/src/routes/blog/hello-world/page.tsx
