+++ packages/fullstack/examples/react-router/src/routes/blog/page.tsx
