+++ packages/fullstack/examples/react-router/src/routes/blog/post.tsx
