+++ packages/fullstack/examples/react-router/src/routes/blog/vite-tips/page.tsx
