--- packages/fullstack/examples/react-router/src/routes/index.css
@@ -1,23 +1,0 @@
-.counter-card {
-  text-align: center;
-}
-
-.counter-card h2 {
-  color: #646cff;
-  margin-bottom: 1rem;
-}
-
-.counter-card p {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin: 1rem 0;
-}
-
-.hero {
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.hero h1 {
-  color: rgb(100, 108, 255);
-}
