+++ packages/fullstack/examples/react-router/src/routes/page.css
