+++ packages/fullstack/examples/vue-router/README.md
@@ -1,3 +1,3 @@
-# Vue Router Custom Framework
+# Vue Router
 
 [![Open in StackBlitz](https://developer.stackblitz.com/img/open_in_stackblitz.svg)](https://stackblitz.com/github/hi-ogawa/vite-plugins/tree/main/packages/fullstack/examples/vue-router)
